+++ client/src/components/History/CurrentHistory/HistoryPanel.vue
@@ -13,7 +13,6 @@ import { useHistoryStore } from "@/stores/historyStore";
 import { useUserStore } from "@/stores/userStore";
 import { type Alias, getOperatorForAlias } from "@/utils/filtering";
 import { setItemDragstart } from "@/utils/setDrag";
-import { startWatchingHistory } from "@/watch/watchHistoryProvided";
 
 import { useHistoryDragDrop } from "../../../composables/historyDragDrop";
 
@@ -329,7 +328,7 @@ function updateContentStats() {
 }
 
 function reloadContents() {
-    startWatchingHistory();
+    historyStore.startWatchingHistory();
 }
 
 function setInvisible(item: HistoryItemSummary) {
