+++ client/src/components/Tool/ToolForm.vue
@@ -128,7 +128,6 @@ import { useHistoryStore } from "@/stores/historyStore";
 import { useTourStore } from "@/stores/tourStore";
 import { useUserStore } from "@/stores/userStore";
 import { useUserToolsServiceCredentialsStore } from "@/stores/userToolsServiceCredentialsStore";
-import { startWatchingHistory } from "@/watch/watchHistoryProvided";
 
 import ToolRecommendation from "../ToolRecommendation";
 import { getToolFormData, submitJob, updateToolFormData } from "./services";
@@ -310,6 +309,7 @@ export default {
     methods: {
         ...mapActions(useJobStore, ["saveLatestResponse"]),
         ...mapActions(useTourStore, ["setTour"]),
+        ...mapActions(useHistoryStore, ["startWatchingHistory"]),
         emailAllowed(config, user) {
             return config.server_mail_configured && !user.isAnonymous;
         },
@@ -428,7 +428,7 @@ export default {
                     this.submissionRequestFailed = false;
                     this.showExecuting = false;
                     let changeRoute = false;
-                    startWatchingHistory();
+                    this.startWatchingHistory();
                     if (jobResponse.produces_entry_points) {
                         this.showEntryPoints = true;
                         this.entryPoints = jobResponse.jobs;
